Move grid power and current formulas to Grid
Now that there is a microgrid.grid object,
the grid metrics are published there.
